+++ app/views/notifier_mailer/notify_not_added_answer.html.erb
@@ -1,7 +1,7 @@
 <div style="font-family: verdana, sans-serif, arial, helvetica;">
   <h3 style="font-weight: normal;">
     <%= markdown_without_paragraph(
-      t('notifier.notify_not_added_answer.warning')
+      t('notifier.notify_action_not_found.warning')
     ) %>
   </h3>
