+++ config/locales/es.yml
@@ -1143,10 +1143,9 @@
       title: "Nuevo comentario en el informe %{review} [#%{finding_id}]"
       finding_data: "Se ha agregado un comentario al hallazgo **%{finding_code}** del informe **%{review}**"
       finding_link: "Ver hallazgo"
-    notify_not_added_answer:
-      title: "No se ha podido agregar el comentario que usted esta enviando"
-      warning: "No se ha podido agregar el comentario que usted esta enviando ya que no estamos recibiendo el mismo asunto en la respuesta, 
-                por favor veriquelo."
+    notify_action_not_found:
+      title: "No se procesó su correo"
+      warning: "No se tomó acción con el correo recibido."
     notify_new_findings:
       title: "Notificación de hallazgos"
       created_title:
